% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACtion-package.R
\docType{package}
\name{ATACtion}
\alias{ATACtion}
\alias{ATACtion-package}
\title{ATACtion: A multiresolution framework for interactive visualization, clustering, and annotation of single-cell epigenomic profiles}
\description{
Extenstion of ACTIONet package for single-cell epigenomic dataset
}
\section{Usage}{
 

\enumerate{
\item ?reduce.peaks.ACTION to reduce raw count matrix in an SingleCellExperiment objects.
\item ?run.ACTIONet to run ACTIONet on SingleCellExperiment objects.
}
}

\section{Useful links}{


\enumerate{
\item Report bugs at \url{https://github.com/shmohammadi86/ATACtion/issues}
}
}

